Initial version of the doc snippet runability checker.
Not hooked up to a completely packaged script yet, run:
./node_modules/.bin/ts-node build-scripts/doc-gen/test-api-snippets.ts source/_data/api/0.0.1/tfjs-layers.json

(or replace -layers with -core)

We take the code snippet, wrap it in an async function with a top level error
handler registered.  This is then embedded into a start scope/end scope block.

One of the current messy bits is that exceptions arise at two different places
1) the outer eval call -- this is where most failures show up and is
handled as you would expect.
2) the inner eval call -- errors due to missing browser (we're running headless)
 features (ImageData/performance) show up here.

This is why we have the repeated error reprorting console.warns, both in code
and in the eval'd error handler string.