Allow unpacking of interfaces across package boundaries. (#254)
To do this, we have to put the interfaces (as well as inlined type and doc aliases) to the individual repository JSON files. Then, we do the replacement during the merge step.

This does not yet support namespaces (e.g. `io.`). I regenerated docs for 0.15.1 to show you the diff of this change (very minor).

InputConfig is no longer referenced:
Before:
![image](https://user-images.githubusercontent.com/1100749/53352506-e3671180-38f0-11e9-8c9e-16d1851d152c.png)
After: 
![image](https://user-images.githubusercontent.com/1100749/53352490-dd713080-38f0-11e9-8496-cd513f0741ad.png)

DataType is unpacked:
Before: 
![image](https://user-images.githubusercontent.com/1100749/53352551-f679e180-38f0-11e9-856a-4d24fbe40ff3.png)
After: 
![image](https://user-images.githubusercontent.com/1100749/53352569-042f6700-38f1-11e9-9d00-6a0d3a4fa3f6.png)